wip: parallelize them tests!
